Change Decoder and detectors to new raw data format
Note that the raw data values are now no longer copied out of the
event buffer into a member vector of Decoder, but instead are
retrieved directly from the event buffer, which is mapped to the event
structure defined in rawdata.h.

We don't do it quite like this in the real analyzer because it
pre-parses the buffer in terms of ROC/module/channel, but in general
the raw event buffer will need to be available during the entire
processing time of an event, so the buffer handling here is a good
test.