ThreadTest: Avoid random number baggage in AnalysisWorker
AnalysisWorker could now actually be a free function instead of a class,
but let's leave it as it as for illustration. Also, while this is
cleaner code, it doesn't actually save much memory since there is still
one heavy generator object per thread. Not that that's a real problem,
of course, it's small enough.

Also, eliminate the unused instance of AnalysisWorker in the main
program.